Move all the OpenMP code to C++; Not as easy as it sounds...
See:

- https://cran.r-project.org/doc/manuals/r-release/R-exts.html#OpenMP-support
- https://stackoverflow.com/questions/54056594/cran-acceptable-way-of-linking-to-openmp-some-c-code-called-from-rcpp
- Related issue: #239
