+++ src/Makevars.in
@@ -8,18 +8,18 @@ CXX_STD = CXX14
 @CXX14STD@
 ################################################################################
 ## Package library flags.
-PKG_LIBS    = $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS) $(SHLIB_OPENMP_CFLAGS)
+PKG_LIBS    = $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS) $(SHLIB_OPENMP_CXXFLAGS)
 
 # Release options
-PKG_CFLAGS  = $(SHLIB_OPENMP_CFLAGS) -D_isRxODE_
-PKG_CXXFLAGS = -D_isRxODE_ -DBOOST_DISABLE_ASSERTS -@ISYSTEM@"$(BH)" -@ISYSTEM@"$(EG)" -@ISYSTEM@"$(SH)" -@ISYSTEM@"$(RCPP)" -@ISYSTEM@"$(ARMA)"
+PKG_CFLAGS  =  -D_isRxODE_
+PKG_CXXFLAGS = $(SHLIB_OPENMP_CXXFLAGS) -D_isRxODE_ -DBOOST_DISABLE_ASSERTS -@ISYSTEM@"$(BH)" -@ISYSTEM@"$(EG)" -@ISYSTEM@"$(SH)" -@ISYSTEM@"$(RCPP)" -@ISYSTEM@"$(ARMA)"
 
 
-SOURCES_C = forder.c lincmt.c call_dvode.c dop853.c tran.c omegaChol.c init.c par_solve.c cfode.c common.c corfailure.c correction.c daxpy.c ddot.c dgefa.c dgesl.c dscal.c fnorm.c idamax.c intdy.c lsoda.c methodswitch.c orderswitch.c prja.c scaleh.c solsy.c stoda.c vmnorm.c strdup_printf.c rprintf.c box.c matexp_HM98.c utilc.c checkmate_stub.c
-SOURCES_CPP = et.cpp boost.cpp fast_factor.cpp etTran.cpp rxData.cpp cvPost.cpp expandGrid.cpp lincmtB.cpp RcppExports.cpp WinDrive.cpp rxInv.cpp rxStack.cpp rxOptRep.cpp expm.cpp threefry.cpp vandercorput.cpp checkmate.cpp
+SOURCES_C =  lincmt.c call_dvode.c dop853.c tran.c omegaChol.c init.c cfode.c common.c corfailure.c correction.c daxpy.c ddot.c dgefa.c dgesl.c dscal.c fnorm.c idamax.c intdy.c lsoda.c methodswitch.c orderswitch.c prja.c scaleh.c solsy.c stoda.c vmnorm.c strdup_printf.c rprintf.c box.c matexp_HM98.c utilc.c checkmate_stub.c
+SOURCES_CPP = par_solve.cpp forder.cpp et.cpp boost.cpp fast_factor.cpp etTran.cpp rxData.cpp cvPost.cpp expandGrid.cpp lincmtB.cpp RcppExports.cpp WinDrive.cpp rxInv.cpp rxStack.cpp rxOptRep.cpp expm.cpp threefry.cpp vandercorput.cpp checkmate.cpp 
 SOURCES_FORTAN = dlsoda.f opkda2.f matexp.f dsphiv.f dgpadm.f
 
-OBJECTS = $(SOURCES_C:.c=.o) $(SOURCES_CPP:.cpp=.o) $(SOURCES_FORTAN:.f=.o) 
+OBJECTS = $(SOURCES_CPP:.cpp=.o) $(SOURCES_C:.c=.o) $(SOURCES_FORTAN:.f=.o) 
 
 # strippedLib: $(SHLIB)
 # 	if test -e "/usr/bin/strip"; then /usr/bin/strip -s $(SHLIB); fi
